requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Data processing
pandas==2.1.3
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3

# YAML configuration
PyYAML==6.0.1

# Telegram Bot
python-telegram-bot==20.7

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0

# Async support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Data validation
email-validator==2.1.0.post1
phonenumbers==8.13.25

# Cryptography and security
bcrypt==4.1.1
passlib[bcrypt]==1.7.4

# Utilities
python-slugify==8.0.1
pillow==10.1.0
openpyxl==3.1.2

# Monitoring and logging
structlog==23.2.0